Move filesourceEntrypointFiles out of CntlrCmdLine.
